+++ metadata.rb
@@ -5,7 +5,7 @@ license 'Apache-2.0'
 description 'Manages organization-wide attributes'
 long_description 'Manages organization-wide attributes via data bags'
 
-version '0.1.0'
+version '0.2.0'
 chef_version '>= 14.0'
 
 source_url 'https://github.com/mattray/organization-attributes-cookbook'